requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
executing==2.0.1
fastapi==0.110.1
h11==0.14.0
httptools==0.6.1
icecream==2.1.3
idna==3.6
motor==3.4.0
numpy==1.26.4
pandas==2.2.1
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic_core==2.16.3
Pygments==2.17.2
pymongo==4.6.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.10.0
tzdata==2024.1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
xlrd==2.0.1
